\name{assortnet-package}
\alias{assortnet-package}
\alias{assortnet}
\docType{package}
\title{
Calculate the assortativity coefficient of weighted and binary networks
~~ assortnet ~~
}
\description{
Functions to calculate the assortment of vertices in social networks.  This can be measured on both weighted and binary networks, with discrete or continuous vertex values.
}
\details{
\tabular{ll}{
Package: \tab assortnet\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2013-05-22\cr
License: \tab GPL2\cr
}
}
\author{
Maintainer: Damien Farine  <dfarine@orn.mpg.de>
}
\references{
	Newman (2003) Mixing patterns in networks. Physical Review E (67)
	Farine, D.R. (2014) Measuring phenotypic assortment in animal social networks: weighted associations are more robust than binary edges. Animal Behaviour 89: 141-153.
}
\keyword{ package }
